app/backend/requirements.txt
#### Any version change made here should also be made and tested for the enrichment and function apps in /functions and /app/enrichment

azure-identity==1.12.0
Flask==2.3.2
langchain>=0.0.157
azure-mgmt-cognitiveservices==13.5.0
openai==0.27.0
# azure-search-documents==11.4.0b3
azure-search-documents==11.4.0b11
azure-storage-blob==12.16.0
azure-cosmos == 4.3.1
tiktoken == 0.4.0
fastapi == 0.109.1
fastapi-utils == 0.2.1
uvicorn == 0.23.2
numexpr == 2.9.0
langchain-experimental==0.0.49
microsoft-bing-websearch
msrest
azure-common
msrestazure
azure-core